colab notebook prototype
it seems that colab has some interesting interactions with conda ... we use `condacolab` in this protoype to set up our environment. Since we only need to (theoretically) run one script in our conda env, we do so with `conda run -n <env> ...`. This has the downside of only printing all results when the script *completes* execution.